More bug fixes

+++ public_html/project/cart.php
@@ -26,6 +26,10 @@ if(!empty($action)) {
             break;
         case "update":
             $des_quan = (int)se($_POST, "desired_quantity", 0, false);
+            if($des_quan < 0) {
+                flash("Quantity can not be set to negative values", "warning");
+                die(header("Location: $BASE_PATH/cart.php"));
+            }
             if($des_quan == 0){
                 $query = "DELETE FROM Cart WHERE id = :cid AND user_id = :uid";
                 $stmt = $db->prepare($query);
@@ -118,7 +122,7 @@ try {
                     <form method="POST">
                         <input type="hidden" name="cart_id" value="<?php se($c, "id"); ?>" />
                         <input type="hidden" name="action" value="update" />
-                        <input type="number" name="desired_quantity" value="<?php se($c, "desired_quantity"); ?>" min="0" max="<?php se($c, "stock"); ?>" />
+                        <input type="number" name="desired_quantity" value="<?php se($c, "desired_quantity"); ?>" max="<?php se($c, "stock"); ?>" />
                         <input type="submit" class="btn btn-primary" value="Update Quantity" />
                     </form>
                 </td>
